requirements.txt
streamlit
spacy
python-docx
PyPDF2
langdetect
reportlab
